refactor: rename subscriber to kiroku
